int to long

+++ inject-generator/src/test/java/io/avaje/inject/generator/models/valid/provider/FactoryProvider.java
@@ -25,7 +25,7 @@ public class FactoryProvider {
 
   @Bean
   @Prototype
-  Provider<Integer> random() {
-    return new Random()::nextInt;
+  Provider<Long> random() {
+    return new Random()::nextLong;
   }
 }
